Added value attribute to area
Added value attribute to area class.
Added Set Value option to right click menu.
Added value attribute to html output.
Adding this as I want an easier way to store a value that is separate from the title and href values, so that I can use the image map as a single or multi-select in a form.
Added additional optional buttons to the quicksettings panel relevant for when embedding the tool in another app.
Updated readme

#TODO: Ensure generated html can be properly exported to parent.